allow alternate multiplexing protocols
