+++ v2/listener.go
@@ -6,8 +6,6 @@ import (
 	"sync"
 	"sync/atomic"
 	"time"
-
-	"github.com/xtaci/smux"
 )
 
 var (
@@ -15,16 +13,15 @@ var (
 )
 
 type ListenOpts struct {
-	Listener          net.Listener
-	BufferSize        int
-	KeepAliveInterval time.Duration
+	Listener net.Listener
+	Protocol Protocol
 }
 
 type listener struct {
 	ListenOpts
 	nextConn              chan net.Conn
 	nextErr               chan error
-	sessions              map[int]*smux.Session
+	sessions              map[int]Session
 	nextSessionID         int
 	closeOnce             sync.Once
 	chClosed              chan struct{}
@@ -36,14 +33,14 @@ type listener struct {
 // Listen creates a net.Listener that multiplexes connections over a connection
 // obtained from the underlying opts.Listener.
 func Listen(opts *ListenOpts) net.Listener {
-	if opts.BufferSize <= 0 {
-		opts.BufferSize = defaultBufferSize
+	if opts.Protocol == nil {
+		opts.Protocol = defaultProtocol
 	}
 	l := &listener{
 		ListenOpts: *opts,
 		nextConn:   make(chan net.Conn, 1000),
 		nextErr:    make(chan error, 1),
-		sessions:   make(map[int]*smux.Session),
+		sessions:   make(map[int]Session),
 		chClosed:   make(chan struct{}),
 	}
 	go l.listen()
@@ -65,12 +62,7 @@ func (l *listener) listen() {
 }
 
 func (l *listener) handleConn(conn net.Conn) {
-	smuxConfig := smux.DefaultConfig()
-	smuxConfig.MaxReceiveBuffer = l.BufferSize
-	if l.KeepAliveInterval > 0 {
-		smuxConfig.KeepAliveInterval = l.KeepAliveInterval
-	}
-	session, err := smux.Server(conn, smuxConfig)
+	session, err := l.Protocol.Server(conn, &l.ListenOpts)
 	if err != nil {
 		l.nextErr <- err
 		return
@@ -98,8 +90,9 @@ func (l *listener) handleConn(conn net.Conn) {
 		}
 		atomic.AddInt64(&l.numVirtualConnections, 1)
 		l.nextConn <- &cmconn{
-			Conn:    stream,
-			onClose: l.cmconnClosed,
+			Conn:            stream,
+			onClose:         l.cmconnClosed,
+			translateErrors: l.Protocol.ErrorMapper(),
 		}
 	}
 }
